requirements.txt
amqp==2.5.2
asgiref==3.2.7
billiard==3.6.3.0
celery==4.4.2
certifi==2020.4.5.1
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==3.0.5
django-environ==0.4.5
django-filter==2.2.0
django-heroku==0.3.1
django-rest-swagger==2.2.0
django-widget-tweaks==1.4.8
djangorestframework==3.11.0
gunicorn==20.0.4
idna==2.9
importlib-metadata==1.6.0
itypes==1.2.0
Jinja2==2.11.2
kombu==4.6.8
Markdown==3.2.1
MarkupSafe==1.1.1
openapi-codec==1.3.2
Pillow==7.1.1
psycopg2==2.8.5
python-http-client==3.2.7
pytz==2019.3
redis==3.5.0
requests==2.23.0
sendgrid==6.3.0
simplejson==3.17.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.9
vine==1.3.0
whitenoise==5.0.1
zipp==3.1.0
